Relocate tests
unit tests -> test/unit
e2e_test -> test/e2e
