+++ .github/workflows/playwright.yml
@@ -15,14 +15,14 @@ jobs:
       with:
         node-version: lts/*
     - name: Install dependencies
-      run: cd e2e_test && npm ci
+      run: cd test/e2e && npm ci
     - name: Install Playwright Browsers
-      run: cd e2e_test && npx playwright install --with-deps
+      run: cd test/e2e && npx playwright install --with-deps
     - name: Run Playwright tests
-      run: cd e2e_test && npm run test:all
+      run: cd test/e2e && npm run test:all
     - uses: actions/upload-artifact@v4
       if: always()
       with:
         name: playwright-report
-        path: e2e_test/playwright-report/
+        path: test/e2e/playwright-report/
         retention-days: 30
