+++ test/e2e/tests/editor.spec.js
