+++ test/unit/blocks/edit/proseCollab.test.js
@@ -2,10 +2,10 @@ import { expect } from '@esm-bundle/chai';
 
 // This is needed to make a dynamic import work that is indirectly referenced
 // from edit/prose/index.js
-const { setNx } = await import('../../../scripts/utils.js');
+const { setNx } = await import('../../../../scripts/utils.js');
 setNx('/bheuaark/', { hostname: 'localhost' });
 
-const pi = await import('../../../blocks/edit/prose/index.js');
+const pi = await import('../../../../blocks/edit/prose/index.js');
 
 describe('Prose collab', () => {
   it('Test awareness status', () => {
