+++ test/unit/milo.js
@@ -1,4 +1,4 @@
-import { setNx, codeBase } from '../scripts/utils.js';
+import { setNx, codeBase } from '../../scripts/utils.js';
 
 const nx = setNx('https://da.live/nx');
 const { setConfig } = await import(`${nx}/scripts/nexter.js`);
